// src/api/generated/models/stage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aegis Backend API
 * OpenAPI spec version: 0.1.0
 */
import type { StageKind } from './stageKind.ts';

export interface Stage {
  /** @nullable */
  closed_at?: string | null;
  closes_at: string;
  id: string;
  kind: StageKind;
  latitude: number;
  longitude: number;
  opens_at: string;
  ordinal: number;
  radius_m: number;
}